Return a string containing the binary sid structure
Previously, the `string_to_sid` method was returning a string containing
the address of the sid structure, possibly truncated due to the
`String#strip` call. For example, if the address was "\001\002\003\000",
then the resulting string would be "\001\002\003"

Since the `string_to_sid` method returned the address of the sid
structure, passing that to `SID.new` would fail, as it expects a string
containing the binary sid structure.